fix(server): treat recorded success as run evidence; move feed-health test to unit suite
