--- examples/personal-agent/portfolio/index.ts
@@ -1,9 +1,0 @@
-/**
- * Live-portfolio path: world-model entities, provider-neutral quote interface,
- * deterministic fixtures, and read-time portfolio calculation over Midas
- * holding events. Lives in the personal-agent example next to the Midas
- * connector — it is domain logic, not a platform contract.
- */
-export * from "./world-model";
-export * from "./quote-provider";
-export * from "./portfolio";
